+++ sanic_jwt/__init__.py
@@ -1,14 +1,18 @@
-__version__ = "1.1.4"
+__version__ = "1.2.0"
 __author__ = "Adam Hopkins"
 __credits__ = "Richard Kuesters"
 
 import logging
 
 from .authentication import Authentication
+from .claim import Claim
 from .configuration import Configuration
-from .decorators import protected, scoped
+from .decorators import inject_user
+from .decorators import protected
+from .decorators import scoped
 from .endpoints import BaseEndpoint
-from .initialization import Initialize, initialize
+from .initialization import Initialize
+from .initialization import initialize
 from .responses import Responses
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
@@ -16,11 +20,13 @@ logging.getLogger(__name__).addHandler(logging.NullHandler())
 
 __all__ = [
     "Authentication",
-    "Initialize",
-    "Configuration",
-    "Responses",
     "BaseEndpoint",
+    "Claim",
+    "Configuration",
     "initialize",
+    "Initialize",
+    "inject_user",
     "protected",
+    "Responses",
     "scoped",
 ]