+++ sanic_jwt/initialization.py
@@ -1,15 +1,22 @@
 import inspect
+
 from collections import namedtuple
 
-from sanic import Blueprint, Sanic
+from sanic import Blueprint
+from sanic import Sanic
 
-from sanic_jwt import endpoints, exceptions
+from sanic_jwt import endpoints
+from sanic_jwt import exceptions
 from sanic_jwt.authentication import Authentication
 from sanic_jwt.configuration import Configuration
-from sanic_jwt.decorators import protected, scoped, inject_user
+from sanic_jwt.decorators import inject_user
+from sanic_jwt.decorators import protected
+from sanic_jwt.decorators import scoped
 from sanic_jwt.responses import Responses
 
-_Handler = namedtuple("_Handler", ["name", "keys", "exception"])
+_Handler = namedtuple(
+    "_Handler", ["name", "keys", "exception", "outside_auth_mode"]
+)
 _EndpointMapping = namedtuple("_EndpointMapping", ["cls", "endpoint", "keys"])
 
 
@@ -40,28 +47,34 @@ endpoint_mappings = (
     ),
 )
 
-handlers = (
-    _Handler("authenticate", None, exceptions.AuthenticateNotImplemented()),
+auth_mode_handlers = (
+    _Handler(
+        "authenticate", None, exceptions.AuthenticateNotImplemented(), False
+    ),
     _Handler(
         "store_refresh_token",
         ["refresh_token_enabled"],
         exceptions.RefreshTokenNotImplemented(),
+        False,
     ),
     _Handler(
         "retrieve_refresh_token",
         ["refresh_token_enabled"],
         exceptions.RefreshTokenNotImplemented(),
+        False,
     ),
-    _Handler("retrieve_user", None, None),
     _Handler(
         "add_scopes_to_payload",
         ["scopes_enabled"],
         exceptions.ScopesNotImplemented(),
+        False,
     ),
-    _Handler("override_scope_validator", None, None),
-    _Handler("destructure_scopes", None, None),
-    _Handler("extend_payload", None, None),
+    _Handler("override_scope_validator", None, None, False),
+    _Handler("destructure_scopes", None, None, False),
+    _Handler("extend_payload", None, None, False),
 )
+auth_mode_agnostic_handlers = (_Handler("retrieve_user", None, None, True),)
+handlers = auth_mode_handlers + auth_mode_agnostic_handlers
 
 init_classes = (
     "configuration_class", "authentication_class", "responses_class"
@@ -102,6 +115,7 @@ class Initialize:
         self.__add_class_views()
         self.__add_endpoints()
         self.__initialize_instance()
+        self.__initialize_claims()
 
     def __check_deprecated(self):
         """
@@ -175,7 +189,7 @@ class Initialize:
         and / or `Responses`) have been overwitten and if they're still valid
         """
         # msg took from BaseAuthentication
-        msg = "Sanic JWT was not initialized properly. It did not " "received an instance of {}"
+        msg = "Sanic JWT was not initialized properly. It did not received " "an instance of {}"
         if not issubclass(self.authentication_class, Authentication):
             raise exceptions.InitializationFailure(
                 message=msg.format("Authentication")
@@ -200,25 +214,34 @@ class Initialize:
         # Initialize instance of the Authentication class
         self.instance.auth = self.authentication_class(self.app, config=config)
 
-        if config.auth_mode():
-            # check if kwargs methods contains authentication methods or if
-            # the authentication auth already has them (if subclassed)
+        init_handlers = handlers if config.auth_mode() else auth_mode_agnostic_handlers
 
-            for handler in handlers:
-                if handler.keys is None:
+        for handler in init_handlers:
+            if handler.keys is None:
+                self.__check_method_in_auth(handler.name, handler.exception)
+            else:
+                if all(map(config.get, handler.keys)):
                     self.__check_method_in_auth(
                         handler.name, handler.exception
                     )
-                else:
-                    if all(map(config.get, handler.keys)):
-                        self.__check_method_in_auth(
-                            handler.name, handler.exception
-                        )
 
-            for handler in handlers:
-                if handler.name in self.kwargs:
-                    method = self.kwargs.pop(handler.name)
-                    setattr(self.instance.auth, handler.name, method)
+        for handler in init_handlers:
+            if handler.name in self.kwargs:
+                method = self.kwargs.pop(handler.name)
+                setattr(self.instance.auth, handler.name, method)
+
+    def __initialize_claims(self):
+        if "extra_verifications" in self.kwargs:
+            self.instance.auth._extra_verifications = self.kwargs.get(
+                "extra_verifications"
+            )
+
+        if "custom_claims" in self.kwargs:
+            try:
+                for claim in self.kwargs.get("custom_claims"):
+                    claim._register(self)
+            except AttributeError:
+                raise exceptions.InvalidCustomClaim
 
     def __check_method_in_auth(self, method_name, exc):
         if method_name not in self.kwargs:
@@ -274,7 +297,9 @@ class Initialize:
             )
 
     def _get_url_prefix(self):
-        bp_url_prefix = self.bp.url_prefix if self.bp.url_prefix is not None else ""
+        bp_url_prefix = (
+            self.bp.url_prefix if self.bp.url_prefix is not None else ""
+        )
         config_url_prefix = self.config.url_prefix()
         url_prefix = bp_url_prefix + config_url_prefix
         return url_prefix