+++ tests/test_endpoints_init_on_bp_multiple.py
@@ -51,10 +51,10 @@ def test_protected_blueprints():
 
     assert response1.status == 401
     assert response1.json.get("exception") == "Unauthorized"
-    assert "Authorization header not present." in response1.json.get('reasons')
+    assert "Authorization header not present." in response1.json.get("reasons")
     assert response2.status == 401
     assert response2.json.get("exception") == "Unauthorized"
-    assert "Authorization cookie not present." in response2.json.get('reasons')
+    assert "Authorization cookie not present." in response2.json.get("reasons")
 
     _, response1 = app.test_client.post(
         "/test1/auth", json={"username": "user1", "password": "abcxyz"}
@@ -111,10 +111,10 @@ def test_protected_blueprints():
 
     assert response1.status == 401
     assert response1.json.get("exception") == "Unauthorized"
-    assert "Auth required." in response1.json.get('reasons')
+    assert "Auth required." in response1.json.get("reasons")
     assert response2.status == 401
     assert response2.json.get("exception") == "Unauthorized"
-    assert "Auth required." in response2.json.get('reasons')
+    assert "Auth required." in response2.json.get("reasons")
 
 
 def test_protected_blueprints_debug():
@@ -146,7 +146,7 @@ def test_protected_blueprints_debug():
 
     assert response1.status == 400
     assert response1.json.get("exception") == "Unauthorized"
-    assert "Signature verification failed." in response1.json.get('reasons')
+    assert "Signature verification failed." in response1.json.get("reasons")
     assert response2.status == 400
     assert response2.json.get("exception") == "Unauthorized"
-    assert "Signature verification failed." in response2.json.get('reasons')
+    assert "Signature verification failed." in response2.json.get("reasons")