+++ tests/test_endpoints_scoped.py
@@ -252,86 +252,100 @@ class TestEndpointsSync(object):
         _, response = sanic_app.test_client.get("/auth/me")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/1")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/2")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/3")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/4")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/5")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/6/1")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/6/foo")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/7/1")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/7/foo")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/8")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/9/1")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
         _, response = sanic_app.test_client.get("/protected/scoped/9/foo")
         assert response.status == 401
         assert response.json.get("exception") == "Unauthorized"
-        assert "Authorization header not present." in \
-            response.json.get('reasons')
+        assert "Authorization header not present." in response.json.get(
+            "reasons"
+        )
 
     def test_scopes_user1(self, app_with_scopes, user1):
         sanic_app, sanicjwt = app_with_scopes
@@ -387,7 +401,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/4",
@@ -434,7 +448,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/8",
@@ -452,7 +466,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
     def test_scopes_user2(self, app_with_scopes, user2):
         sanic_app, sanicjwt = app_with_scopes
@@ -626,7 +640,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/4",
@@ -635,7 +649,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/5",
@@ -644,7 +658,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/1",
@@ -653,7 +667,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/foo",
@@ -662,7 +676,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/7/1",
@@ -671,7 +685,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/8",
@@ -689,7 +703,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
     def test_scopes_user4(self, app_with_scopes, user4):
         sanic_app, sanicjwt = app_with_scopes
@@ -727,7 +741,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/2",
@@ -736,7 +750,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/3",
@@ -745,7 +759,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/4",
@@ -754,7 +768,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/5",
@@ -763,7 +777,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/1",
@@ -772,7 +786,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/foo",
@@ -781,7 +795,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/7/1",
@@ -800,7 +814,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/8",
@@ -809,7 +823,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/9/1",
@@ -828,7 +842,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
     def test_scopes_user5(self, app_with_scopes, user5):
         sanic_app, sanicjwt = app_with_scopes
@@ -866,7 +880,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/2",
@@ -875,7 +889,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/3",
@@ -884,7 +898,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/4",
@@ -902,7 +916,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/1",
@@ -911,7 +925,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/foo",
@@ -920,7 +934,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/7/1",
@@ -929,7 +943,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/8",
@@ -947,7 +961,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
     def test_scopes_user6(self, app_with_scopes, user6):
         sanic_app, sanicjwt = app_with_scopes
@@ -985,7 +999,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/2",
@@ -994,7 +1008,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/3",
@@ -1003,7 +1017,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/4",
@@ -1012,7 +1026,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/5",
@@ -1021,7 +1035,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/1",
@@ -1030,7 +1044,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/6/foo",
@@ -1039,7 +1053,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/7/1",
@@ -1048,7 +1062,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/8",
@@ -1057,7 +1071,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
         _, response = sanic_app.test_client.get(
             "/protected/scoped/9/1",
@@ -1066,7 +1080,7 @@ class TestEndpointsSync(object):
 
         assert response.status == 403
         assert response.json.get("exception") == "Unauthorized"
-        assert "Invalid scope." in response.json.get('reasons')
+        assert "Invalid scope." in response.json.get("reasons")
 
 
 def test_no_user_scopes(app_with_scopes):
@@ -1090,7 +1104,7 @@ def test_no_user_scopes(app_with_scopes):
 
     assert response.status == 403
     assert response.json.get("exception") == "Unauthorized"
-    assert "Invalid scope." in response.json.get('reasons')
+    assert "Invalid scope." in response.json.get("reasons")
 
 
 def test_scoped_option(app_with_scopes):
@@ -1163,7 +1177,7 @@ def test_scoped_with_override(app_with_scopes_override):
 
     assert response.status == 403
     assert response.json.get("exception") == "Unauthorized"
-    assert "Invalid scope." in response.json.get('reasons')
+    assert "Invalid scope." in response.json.get("reasons")
 
 
 def test_scoped_with_destructure(app_with_scopes_destructure):