+++ tests/test_endpoints_sync_methods.py
@@ -152,10 +152,9 @@ class TestEndpointsSync(object):
                 headers={"Authorization": "Bearer {}".format(access_token)},
             )
 
-            assert response.status == 403
+            assert response.status == 401
             assert response.json.get("exception") == "InvalidToken"
-            assert "Signature has expired." in \
-                response.json.get('reasons')
+            assert "Signature has expired." in response.json.get("reasons")
 
     def test_refresh_token_sync(
         self, app_with_sync_methods, authenticated_response